proyecto_acceso_biometrico/requirements.txt
psycopg2-binary
opencv-contrib-python
python-dotenv
Pillow